Check for validity by state, also returns requirements by state
This includes a small refactor that eliminates some duplication, but more importantly adds two new functions.

The first 'validateForState' a boolean if the dl is valid for a specific state/states, this provides a bit of an alternative interface to this package.

Then I added 'requirementsByState' that returns the dl requirements for a specific state, this way we can return validation messages that let folks more easily fix drivers license inputs.